+++ contents/ui/configAppearance.qml
@@ -75,6 +75,8 @@ KCM.SimpleKCM {
     property int cfg_retryMaxSecondsDefault: 300
     property string cfg_defaultSorting: "status"
     property string cfg_defaultSortingDefault: "status"
+    property var cfg_customGuestOrder: []
+    property var cfg_customGuestOrderDefault: []
     property string cfg_compactMode: "cpu"
     property string cfg_compactModeDefault: "cpu"
     property bool cfg_enableNotifications: true